+++ core/CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump protocol version to 16.
 - `StructSerializer::serialize_field()` takes `id` now as `impl Into<u32>`.
 - `Serializer::serialize_enum()` takes `variant` now as `impl Into<u32>`.
+- Replace `VecDeserializer::remaining_elements()` and `has_more_elements()` with `len()` and
+  `is_empty()`.
+- Replace `MapDeserializer::remaining_elements()` and `has_more_elements()` with `len()` and
+  `is_empty()`.
+- Replace `SetDeserializer::remaining_elements()` and `has_more_elements()` with `len()` and
+  `is_empty()`.
 
 ## [0.4.0] - 2024-03-21